+++ aa_discord_announcements/constants.py
@@ -6,3 +6,6 @@ APP_NAME = "aa-discord-announcements"
 GITHUB_URL = f"https://github.com/ppfeufer/{APP_NAME}"
 
 DISCORD_WEBHOOK_REGEX = r"https:\/\/discord\.com\/api\/webhooks\/[\d]+\/[a-zA-Z0-9_-]+$"
+
+# All internal URLs need to start with this prefix
+INTERNAL_URL_PREFIX = "-"